project page and form modified with mantine and color palette

+++ src/pages/Project.jsx
@@ -4,6 +4,7 @@ import { Link, useNavigate, useParams } from "react-router-dom";
 import { AuthContext } from "../contexts/Auth.context";
 import { BASE_URL } from "../config/config.index";
 import Comments from "../components/Comments";
+import { Button} from "@mantine/core";
 import "./Project.css"
 
 function Project() {
@@ -33,31 +34,34 @@ function Project() {
 
    //Deleting the project
    const handleDeleteButton = async () => {
-      try {
-         //Deleting all comments related to the project first when there are comments
-         if (allComments.length > 0) {
-            allComments.forEach(async (comment) => {
-               try {
-                  const responseComments = await axios.delete(
-                     `${BASE_URL}/project/${projectId}/comments/delete`
-                  );
-                  if (responseComments.status === 204) {
-                     console.log("comments deleted");
+      const confirmDelete = window.confirm("Are you sure you want to delete this comment?");
+      if (confirmDelete) {
+         try {
+            //Deleting all comments related to the project first when there are comments
+            if (allComments.length > 0) {
+               allComments.forEach(async (comment) => {
+                  try {
+                     const responseComments = await axios.delete(
+                        `${BASE_URL}/project/${projectId}/comments/delete`
+                     );
+                     if (responseComments.status === 204) {
+                        console.log("comments deleted");
+                     }
+                  } catch (error) {
+                     console.error(error);
                   }
-               } catch (error) {
-                  console.error(error);
-               }
-            });
+               });
+            }
+            //After deleting all comments, then delete the current project
+            const response = await axios.delete(
+               `${BASE_URL}/project/delete/${projectId}`
+            );
+            if (response.status === 200) {
+               navigate(`/user/${user._id}`);
+            }
+         } catch (error) {
+            console.error(error);
          }
-         //After deleting all comments, then delete the current project
-         const response = await axios.delete(
-            `${BASE_URL}/project/delete/${projectId}`
-         );
-         if (response.status === 200) {
-            navigate(`/user/${user._id}`);
-         }
-      } catch (error) {
-         console.error(error);
       }
    }
 
@@ -82,8 +86,8 @@ function Project() {
                <h1>Project : {title}</h1>
                {user && userId._id === user._id && (
                <div className="project_buttons">
-                  <button className="project_link" onClick={handleUpdateButton}>Update Project</button>
-                  <button className="project_link" onClick={handleDeleteButton}>Delete Project</button>
+                  <Button className="project_link" onClick={handleUpdateButton}>Update Project</Button>
+                  <Button className="project_link" onClick={handleDeleteButton}>Delete Project</Button>
                </div>
             )}
          </div>
@@ -97,8 +101,8 @@ function Project() {
                </ul>
                <div className="project_second_part">
                <p className="description">{description}</p>
-               <Link className="project_links" to={repositoryLink}>Link to repo</Link>
-               <Link className="project_links" to={projectFolder}>Download project</Link>
+               <a className="project_links" href={repositoryLink} target="blank">Link to repo</a>
+               <a className="project_links" href={projectFolder} target="blank">Download project</a>
                <p className="creator">by <Link className="author" to={`/user/${userId._id}`}>{userId.username}</Link></p>
                </div>
             </div>
